Speed up lookups in AppDomainInfo class
Profiling has shown that the previously used immutable dictionaries are much slower than regular dictionaries. After populating the dictionaries with type information the dictionaries are static with the rare exception that an assembly is loaded dynamically lateron. In this case the dictionaries are copied, modified and replaced atomically. Users of the the AppDomainInfo class can treat the dictionaries as immutable.