Revise integration of external object serializers
External object serializers must derive from the ExternalObjectSerializer<T> class now. The generic type argument T specifies the type the serializer can handle. The type can be a concrete type, a closed generic type, a generic type definition or an interface. Generic interfaces are supported as well. The ExternalObjectSerializer attribute is still attached to an external object serializer implementation, but it does not carry along the serialized type any more.