Update GriffinPlus.Lib.Common to version 3.1.1
The serializer uses the RuntimeMetadata class in the GriffinPlus.Lib.Common package when scanning for custom serializers. This class crashed if an assembly was loaded multiple times into the application domain. On .NET Framework this can occur when an assembly is loaded into the different load contexts. The *Load* context is the usual context for assemblies loaded from the Global Assembly Cache (GAC) and assemblies loaded from the application's base directory or a private bin path below the base directory. Other assemblies may be loaded into the *LoadFrom* context causing the same assembly occur twice in the application domain. On .NET Core 2/3 and .NET 5+ the same issue can happen if assemblies are loaded into different *Assembly Load Contexts*.

The GriffinPlus.Lib.Common package was updated to version 3.1.1 fixing the crash.

The RuntimeMetadata class now supports .NET Framework 4+, .NET Core 2/3 and .NET 5+ properly.

On .NET Framework 4+ the following assemblies are considered:
- Assemblies in the Global Assembly Cache (GAC)
- Assemblies in the application's base directory or a private bin path below the base directory
- Assemblies that have been generated dynamically

On .NET Core 2/3 and .NET 5+ the following assemblies are considered:
- Assemblies that are loaded into the default assembly load context
- Assemblies that have been generated dynamically

These assemblies are scanned for custom serializers now. Other assemblies are not scanned.